requirements.txt
Apache-Airflow
apache-airflow
requests
pytest
setuptools
apache-airflow-providers-apache-livy
apache-airflow-microsoft-fabric-plugin
msal